@@ -4,26 +4,26 @@
   "license": "MIT",
   "private": true,
   "dependencies": {
-    "@babel/polyfill": "^7.2.5",
-    "@material-ui/core": "^3.9.0",
-    "@material-ui/icons": "^3.0.2",
+    "@babel/polyfill": "^7.6.0",
+    "@material-ui/core": "^4.4.2",
+    "@material-ui/icons": "^4.4.1",
     "autosuggest-highlight": "^3.1.1",
     "isomorphic-fetch": "^2.2.1",
     "lodash.debounce": "^4.0.8",
     "moment": "^2.24.0",
-    "moment-timezone": "^0.5.23",
-    "prop-types": "^15.6.2",
-    "rc-slider": "^8.6.4",
-    "react": "^16.7.0",
+    "moment-timezone": "^0.5.26",
+    "prop-types": "^15.7.2",
+    "rc-slider": "^8.7.1",
+    "react": "^16.9.0",
     "react-autosuggest": "^9.4.3",
-    "react-dom": "^16.7.0",
+    "react-dom": "^16.9.0",
     "react-jss": "^8.6.1",
-    "react-redux": "^5.1.1",
-    "react-router-dom": "^4.3.1",
-    "redux": "^4.0.1",
-    "redux-saga": "^0.16.2",
+    "react-redux": "^7.1.1",
+    "react-router-dom": "^5.0.1",
+    "redux": "^4.0.4",
+    "redux-saga": "^1.1.1",
     "redux-thunk": "^2.3.0",
-    "typeface-roboto": "0.0.54",
+    "typeface-roboto": "0.0.75",
     "workbox-cli": "^3.6.3",
     "workbox-sw": "^3.6.3"
   },
@@ -37,7 +37,7 @@
   },
   "proxy": "http://localhost:3001",
   "devDependencies": {
-    "react-scripts": "3.0.1",
+    "react-scripts": "^3.1.2",
     "redux-devtools": "^3.5.0",
     "redux-devtools-dock-monitor": "^1.1.3",
     "redux-devtools-instrument": "^1.9.6",
