@@ -5,16 +5,16 @@
   "main": "build/index.js",
   "scripts": {
     "build": "NODE_ENV=production webpack",
-    "start": "NODE_ENV=production node -r esm build/index",
+    "start": "NODE_ENV=production node -r esm -r dotenv/config build/index",
     "watch": "NODE_ENV=development PORT=3001 concurrently \"webpack --watch\" \"nodemon --watch build/index.js --exec node -r dotenv/config build/index.js\""
   },
   "author": "Vladimir Feskov",
   "license": "MIT",
   "dependencies": {
-    "@babel/core": "^7.2.2",
+    "@babel/core": "^7.6.0",
     "aws-sdk": "^2.395.0",
-    "babel-loader": "^8.0.5",
-    "babel-preset-react-app": "^9.0.0",
+    "babel-loader": "^8.0.6",
+    "babel-preset-react-app": "^9.0.2",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.18.3",
     "compression": "^1.7.3",
@@ -26,7 +26,7 @@
     "mongodb": "^3.1.13",
     "mongoose": "^5.6.8",
     "node-fetch": "2.6.0",
-    "react": "^16.7.0",
+    "react": "^16.9.0",
     "serve-static": "^1.13.2",
     "webpack": "^4.29.0"
   },
