@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React, { useState } from 'react'
 import Tooltip from '@material-ui/core/Tooltip'
 import moment from 'moment-timezone'
 import withStyles from '@material-ui/core/styles/withStyles'
@@ -6,37 +6,32 @@ import HourpickerDialog from './HourpickerDialog'
 import { timeText, fromUTC } from './Hourpicker-util'
 import styles from './Hourpicker-styles'
 
-class Hourpicker extends Component {
-  state = {
-    dialogOpen: false
+function Hourpicker (props) {
+  const [dialogOpen, setDialogOpen] = useState(false)
+
+  const openDialog = () => {
+    setDialogOpen(true)
   }
 
-  openDialog = () => {
-    this.setState({ dialogOpen: true })
+  const closeDialog = () => {
+    setDialogOpen(false)
   }
 
-  closeDialog = () => {
-    this.setState({ dialogOpen: false })
-  }
-
-  render() {
-    const { utcHour, classes, onSave } = this.props
-    const { dialogOpen } = this.state
-    return (
-      <div className={classes.container}>
-        <Tooltip title={`Timezone: ${moment.tz.guess()}`} placement="bottom">
-          <div>&nbsp;at <a className={classes.text} onClick={this.openDialog}>{timeText(fromUTC(utcHour))}</a></div>
-        </Tooltip>
-        <HourpickerDialog
-          classes={classes}
-          open={dialogOpen}
-          onClose={this.closeDialog}
-          onChange={onSave}
-          value={utcHour}
-        />
-      </div>
-    )
-  }
+  const { utcHour, classes, onSave } = props
+  return (
+    <div className={classes.container}>
+      <Tooltip title={`Timezone: ${moment.tz.guess()}`} placement="bottom">
+        <div>&nbsp;at <a className={classes.text} onClick={openDialog}>{timeText(fromUTC(utcHour))}</a></div>
+      </Tooltip>
+      <HourpickerDialog
+        classes={classes}
+        open={dialogOpen}
+        onClose={closeDialog}
+        onChange={onSave}
+        value={utcHour}
+      />
+    </div>
+  )
 }
 
 export default withStyles(styles)(Hourpicker)
