@@ -1,6 +1,6 @@
 import App from '../components/App'
 import { connect } from 'react-redux'
-import { setShowIntro, signOut, unwatch, setStarsOpen } from '../actions'
+import { setShowIntro, signOut, unwatch, setStarsOpen, patchProfile } from '../actions'
 import { withRouter } from 'react-router-dom'
 
 export default withRouter(connect(
@@ -14,6 +14,7 @@ export default withRouter(connect(
     setShowIntro,
     setStarsOpen,
     signOut: signOut.request,
-    unwatch: unwatch.request
+    unwatch: unwatch.request,
+    patchProfileSuccess: patchProfile.success
   }
 )(App))
