@@ -1,18 +1,16 @@
 import React from 'react'
 import { renderToString } from 'react-dom/server'
 import Root from './components/Root'
-import { SheetsRegistry } from 'jss'
-import JssProvider from 'react-jss/lib/JssProvider'
 import MuiThemeProvider from '@material-ui/core/styles/MuiThemeProvider'
-import createGenerateClassName from '@material-ui/core/styles/createGenerateClassName'
 import configureStore from './store/configureStore'
 import rootSaga from './sagas'
 import createTheme from './theme/createTheme'
 import * as api from './services/api'
 import * as cookieSvc from './services/cookie'
 import { StaticRouter } from 'react-router-dom'
+import { ServerStyleSheets } from '@material-ui/styles'
 
-export async function renderToStrings (serverPort, cookie, url) {
+export function renderToStrings (serverPort, cookie, url) {
   api.setBase({
     url: `http://localhost:${serverPort}`,
     opts: { headers: { cookie } }
@@ -20,25 +18,26 @@ export async function renderToStrings (serverPort, cookie, url) {
   cookieSvc.setSource(cookie)
 
   const store = configureStore()
-  const sagaDone = store.runSaga(rootSaga).done
-  store.close()
-  await sagaDone
-  const sheetsRegistry = new SheetsRegistry()
   const theme = createTheme()
-  const generateClassName = createGenerateClassName()
-
+  const sheets = new ServerStyleSheets();
   const routerContext = {}
-  const html = renderToString(
-    <JssProvider registry={sheetsRegistry} generateClassName={generateClassName}>
-      <StaticRouter location={url} context={routerContext}>
-        <MuiThemeProvider theme={theme} sheetsManager={new Map()}>
-          <Root store={store} />
-        </MuiThemeProvider>
-      </StaticRouter>
-    </JssProvider>
+  const rootComp = sheets.collect(
+    <StaticRouter location={url} context={routerContext}>
+      <MuiThemeProvider theme={theme} sheetsManager={new Map()}>
+        <Root store={store} />
+      </MuiThemeProvider>
+    </StaticRouter>
   )
-  const css = sheetsRegistry.toString()
-  const state = JSON.stringify(store.getState())
 
-  return { html, css, state }
+  const stringsPromise = store.runSaga(rootSaga).toPromise().then(() => {
+    const html = renderToString(rootComp)
+    const css = sheets.toString()
+    const state = JSON.stringify(store.getState())
+    return { html, css, state }
+  })
+
+  renderToString(rootComp)
+  store.close()
+
+  return stringsPromise
 }
